+++ include/bbque/application_manager.h
@@ -127,12 +127,12 @@ public:
 	/**
 	 * @see ApplicationManagerConfIF
 	 */
-	ExitCode_t SetGoalGapEXC(AppPtr_t papp, int8_t gap);
+	ExitCode_t SetGoalGapEXC(AppPtr_t papp, int gap);
 
 	/**
 	 * @see ApplicationManagerConfIF
 	 */
-	ExitCode_t SetGoalGapEXC(AppPid_t pid, uint8_t exc_id, int8_t gap);
+	ExitCode_t SetGoalGapEXC(AppPid_t pid, uint8_t exc_id, int gap);
 
 	/**
 	 * @see ApplicationManagerConfIF
