+++ include/bbque/rtlib.h
@@ -519,7 +519,7 @@ typedef RTLIB_ExitCode_t (*RTLIB_ClearConstraints_t)(
  */
 typedef RTLIB_ExitCode_t (*RTLIB_SetGoalGap_t)(
 		RTLIB_ExecutionContextHandler_t ech,
-		int8_t gap);
+		int gap);
 
 /**@}*/
